Move adjustAmountWithUpfrontFee to common utils
